# improve webidl.is.x performance for common case