--- a/lib/web/fetch/formdata.js
+++ b/lib/web/fetch/formdata.js
@@ -194,11 +194,16 @@ class FormData {
   static setFormDataState (formData, newState) {
     formData.#state = newState
   }
+
+  static isFormData (value) {
+    return value !== null && typeof value === 'object' && #state in value
+  }
 }
 
-const { getFormDataState, setFormDataState } = FormData
+const { getFormDataState, setFormDataState, isFormData } = FormData
 Reflect.deleteProperty(FormData, 'getFormDataState')
 Reflect.deleteProperty(FormData, 'setFormDataState')
+Reflect.deleteProperty(FormData, 'isFormData')
 
 iteratorMixin('FormData', FormData, getFormDataState, 'name', 'value')
 
@@ -256,6 +261,6 @@ function makeEntry (name, value, filename) {
   return { name, value }
 }
 
-webidl.is.FormData = webidl.util.MakeTypeAssertion(FormData.prototype)
+webidl.is.FormData = isFormData
 
 module.exports = { FormData, makeEntry, setFormDataState }