--- a/lib/web/fetch/request.js
+++ b/lib/web/fetch/request.js
@@ -864,15 +864,20 @@ class Request {
   static setRequestState (request, newState) {
     request.#state = newState
   }
+
+  static isRequest (value) {
+    return value !== null && typeof value === 'object' && #state in value
+  }
 }
 
-const { setRequestSignal, getRequestDispatcher, setRequestDispatcher, setRequestHeaders, getRequestState, setRequestState } = Request
+const { setRequestSignal, getRequestDispatcher, setRequestDispatcher, setRequestHeaders, getRequestState, setRequestState, isRequest } = Request
 Reflect.deleteProperty(Request, 'setRequestSignal')
 Reflect.deleteProperty(Request, 'getRequestDispatcher')
 Reflect.deleteProperty(Request, 'setRequestDispatcher')
 Reflect.deleteProperty(Request, 'setRequestHeaders')
 Reflect.deleteProperty(Request, 'getRequestState')
 Reflect.deleteProperty(Request, 'setRequestState')
+Reflect.deleteProperty(Request, 'isRequest')
 
 mixinBody(Request, getRequestState)
 
@@ -986,7 +991,7 @@ Object.defineProperties(Request.prototype, {
   }
 })
 
-webidl.is.Request = webidl.util.MakeTypeAssertion(Request.prototype)
+webidl.is.Request = isRequest
 
 // https://fetch.spec.whatwg.org/#requestinfo
 webidl.converters.RequestInfo = function (V, prefix, argument) {