--- a/lib/web/fetch/response.js
+++ b/lib/web/fetch/response.js
@@ -297,13 +297,18 @@ class Response {
   static setResponseState (response, newState) {
     response.#state = newState
   }
+
+  static isResponse (value) {
+    return value !== null && typeof value === 'object' && #state in value
+  }
 }
 
-const { getResponseHeaders, setResponseHeaders, getResponseState, setResponseState } = Response
+const { getResponseHeaders, setResponseHeaders, getResponseState, setResponseState, isResponse } = Response
 Reflect.deleteProperty(Response, 'getResponseHeaders')
 Reflect.deleteProperty(Response, 'setResponseHeaders')
 Reflect.deleteProperty(Response, 'getResponseState')
 Reflect.deleteProperty(Response, 'setResponseState')
+Reflect.deleteProperty(Response, 'isResponse')
 
 mixinBody(Response, getResponseState)
 
@@ -619,7 +624,7 @@ webidl.converters.ResponseInit = webidl.dictionaryConverter([
   }
 ])
 
-webidl.is.Response = webidl.util.MakeTypeAssertion(Response.prototype)
+webidl.is.Response = isResponse
 
 module.exports = {
   isNetworkError,