--- a/lib/web/websocket/stream/websocketerror.js
+++ b/lib/web/websocket/stream/websocketerror.js
@@ -62,10 +62,15 @@ class WebSocketError extends DOMException {
     error.#reason = reason
     return error
   }
+
+  static isWebSocketError (value) {
+    return value !== null && typeof value === 'object' && #closeCode in value
+  }
 }
 
-const { createUnvalidatedWebSocketError } = WebSocketError
-delete WebSocketError.createUnvalidatedWebSocketError
+const { createUnvalidatedWebSocketError, isWebSocketError } = WebSocketError
+Reflect.deleteProperty(WebSocketError, 'createUnvalidatedWebSocketError')
+Reflect.deleteProperty(WebSocketError, 'isWebSocketError')
 
 Object.defineProperties(WebSocketError.prototype, {
   closeCode: kEnumerableProperty,
@@ -78,6 +83,6 @@ Object.defineProperties(WebSocketError.prototype, {
   }
 })
 
-webidl.is.WebSocketError = webidl.util.MakeTypeAssertion(WebSocketError.prototype)
+webidl.is.WebSocketError = isWebSocketError
 
 module.exports = { WebSocketError, createUnvalidatedWebSocketError }